refactor(native): extract browser reconciliation worker
